% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{calculate_BIC}
\alias{calculate_BIC}
\title{Calculate BIC for model selection}
\usage{
calculate_BIC(log_likelihood, n, params, has_uniform)
}
\arguments{
\item{log_likelihood}{Log-likelihood value}

\item{n}{Number of observations}

\item{params}{Model parameters}

\item{has_uniform}{Boolean for uniform component}
}
\value{
BIC value
}
\description{
Calculate BIC for model selection
}
