% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_log_likelihood}
\alias{calculate_log_likelihood}
\title{Calculate log-likelihood}
\usage{
calculate_log_likelihood(data, params, has_uniform = FALSE)
}
\arguments{
\item{data}{Data matrix}

\item{params}{Model parameters}

\item{has_uniform}{Boolean for uniform component}
}
\value{
Log-likelihood value
}
\description{
Calculate log-likelihood
}
