% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{E_step}
\alias{E_step}
\title{E-step: Calculate log-responsibilities}
\usage{
E_step(data, params, has_uniform)
}
\arguments{
\item{data}{Data matrix}

\item{params}{Model parameters}

\item{has_uniform}{Boolean indicating if uniform component is included}
}
\value{
Matrix of log-responsibilities
}
\description{
E-step: Calculate log-responsibilities
}
\keyword{internal}
