% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{Initialize Gaussian Mixture Model parameters}
\usage{
initialize_parameters(data, K)
}
\arguments{
\item{data}{Matrix of data (n x d)}

\item{K}{Number of Gaussian components}
}
\value{
List with initialized parameters
}
\description{
Initialize Gaussian Mixture Model parameters
}
