% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{M_step}
\alias{M_step}
\title{M-step: Update parameters}
\usage{
M_step(data, log_resp, params, has_uniform)
}
\arguments{
\item{data}{Data matrix}

\item{log_resp}{Log-responsibilities from E-step}

\item{params}{Model parameters}

\item{has_uniform}{Boolean indicating if uniform component is included}
}
\value{
Updated parameters
}
\description{
M-step: Update parameters
}
\keyword{internal}
