% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{plot.gmm_comparison}
\alias{plot.gmm_comparison}
\title{Plot method for model comparison}
\usage{
\method{plot}{gmm_comparison}(x, ...)
}
\arguments{
\item{x}{gmm_comparison object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot method for model comparison
}
