% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{print.gmm_comparison}
\alias{print.gmm_comparison}
\title{Print method for model comparison}
\usage{
\method{print}{gmm_comparison}(x, ...)
}
\arguments{
\item{x}{gmm_comparison object}

\item{...}{Additional arguments}
}
\description{
Print method for model comparison
}
