% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_metrics.R
\name{print.gmm_metrics}
\alias{print.gmm_metrics}
\title{Print method for metrics}
\usage{
\method{print}{gmm_metrics}(x, digits = 3, ...)
}
\arguments{
\item{x}{gmm_metrics object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments}
}
\description{
Print method for metrics
}
