% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{run_multiple_initializations}
\alias{run_multiple_initializations}
\title{Run multiple initializations with error handling}
\usage{
run_multiple_initializations(data, K, n_init, has_uniform)
}
\arguments{
\item{data}{Data matrix}

\item{K}{Number of Gaussian components}

\item{n_init}{Number of initializations}

\item{has_uniform}{Boolean indicating if uniform component is included}
}
\value{
Best model from initializations
}
\description{
Run multiple initializations with error handling
}
\keyword{internal}
