% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_best_model}
\alias{select_best_model}
\title{Model selection using BIC}
\usage{
select_best_model(data, K, n_init = 20)

select_best_model(data, K, n_init = 20)
}
\arguments{
\item{data}{Data matrix}

\item{K}{Number of Gaussian components}

\item{n_init}{Number of random initializations}
}
\value{
Best model according to BIC
}
\description{
Model selection using BIC

Model selection using BIC
}
