% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_metrics.R
\name{summary.gmm_metrics}
\alias{summary.gmm_metrics}
\title{Summary method for metrics}
\usage{
\method{summary}{gmm_metrics}(object, ...)
}
\arguments{
\item{object}{gmm_metrics object}

\item{...}{Additional arguments}
}
\description{
Summary method for metrics
}
